pyproject.toml
[project]
name = "quicktune"
version = "0.0.0.dev0"
dependencies = [
  "torch",
  "torchvision",
  "torchaudio",
  "ConfigSpace",
  "gpytorch",
  "pandas",
  "pyyaml",
  "scikit-learn",
  "timm",
  "tqdm",
]
requires-python = ">=3.9"
authors = [{ name = "Ivo Rapant", email = "rapanti@cs.uni-freiburg.de" }]
readme = "README.md"
description = "Quick-Tune: Quickly Learning Which Pretrained Model to Finetune and How"
classifiers = [
  'Intended Audience :: Science/Research',
  'Intended Audience :: Developers',
  'License :: OSI Approved :: BSD License',
  'Programming Language :: Python',
  'Topic :: Software Development',
  'Topic :: Scientific/Engineering',
  'Operating System :: POSIX',
  'Operating System :: Unix',
  'Programming Language :: Python :: 3',
]
license = { file = "LICENSE" }
